backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pyyaml>=6.0.1
python-multipart>=0.0.12
pydantic>=2.10.0

